Matlab doesn't know about int64_t
--CAR

+++ hardware/X6-1000M/src/X6ADC/libx6adc.cpp
@@ -120,7 +120,7 @@ int stop(int deviceID) {
 	return X6s_[deviceID].stop();
 }
 
-int transfer_waveform(int deviceID, int channel, int64_t *buffer, unsigned int bufferLength) {
+int transfer_waveform(int deviceID, int channel, long long *buffer, unsigned int bufferLength) {
 	// convert from MATLAB 1-based indexing to C 0-based indexing
 	return X6s_[deviceID].transfer_waveform(channel-1, buffer, bufferLength);
 }
